mod_11_lesson_1/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
fonttools==4.53.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.9.0
multidict==6.0.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.7.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rawkit==0.6.0
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
